pyproject.toml
[tool.poetry]
name = "uqcsbot"
version = "0.1.0"
description = "UQCSbot is a chat bot built in Python for the UQCS Discord server."
authors = ["UQ Computing Society <contact@uqcs.org.au>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.1.0"
icalendar = "^4.1"
"python-dateutil" = "^2.8"
pytz = "^2022.2"
requests = "^2.28"
beautifulsoup4 = "^4.11"
APScheduler = "^3.9"
SQLAlchemy = {version = "^1.4.40", extras = ["postgresql_psycopg2binary"]}
humanize = "^4.3"
aiohttp = "^3.8"
emoji = "^1.7.0"
pytest-datafiles = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.2.1"]
build-backend = "poetry.core.masonry.api"
